+++ CHANGELOG.md
@@ -9,3 +9,4 @@ v1.0.1
 ------
 * Added default template
 * Added TravisCI support
+* Added CentOS support (compile from source)